docs: adjust wording for question 15, include framweork
This may not be a necessary change, but I feel Angular, Svelte, and Vue are better categorized as frameworks, while React is a library.

We can also consider adding Remix, Solid, and Astro, but they are probably more aimed towards user interfaces instead of game dev. I can see the existing options being used in conjunction with web game dev.